job-matching-calculator/requirements.txt
# autopep8==1.5.5
# certifi==2020.12.5
# chardet==4.0.0
# click==7.1.2
Flask==1.1.2
# idna==2.10
# itsdangerous==1.1.0
# Jinja2==2.11.3
# MarkupSafe==1.1.1
numpy==1.20.0
pandas==1.2.1
# pep8==1.7.1
psycopg2==2.8.6
# pycodestyle==2.6.0
# python-dateutil==2.8.1
python-dotenv==0.15.0
# pytz==2021.1
# requests==2.25.1
# six==1.15.0
# toml==0.10.2
# urllib3==1.26.3
# Werkzeug==1.0.1
